@@ -16,51 +16,126 @@ import {UnregisteredMessage} from './Messages/UnregisteredMessage';
 import {EventMessage} from './Messages/EventMessage';
 import {HelloMessage} from './Messages/HelloMessage';
 import {AbortMessage} from './Messages/AbortMessage';
-import {ReplaySubject} from 'rxjs/ReplaySubject';
 import {IMessage} from './Messages/Message';
 import {Utils} from './Common/Utils';
-import {Observable} from 'rxjs/Observable';
-import {Subscription} from 'rxjs/Subscription';
-import {Scheduler, Subject} from 'rxjs';
+import {empty, Observable, of, ReplaySubject, Subject, Subscription, throwError, timer, asyncScheduler} from 'rxjs';
 
-import 'rxjs/add/operator/filter';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/mergeMap';
-import 'rxjs/add/operator/merge';
-import 'rxjs/add/operator/take';
-import 'rxjs/add/operator/share';
-import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/mapTo';
-import 'rxjs/add/operator/share';
-import 'rxjs/add/operator/retryWhen';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/switchMapTo';
-import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/takeWhile';
-import 'rxjs/add/operator/delay';
-import 'rxjs/add/operator/publishReplay';
-import 'rxjs/add/operator/publish';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/finally';
-import 'rxjs/add/operator/exhaust';
-import 'rxjs/add/operator/defaultIfEmpty';
-import 'rxjs/add/operator/multicast';
-import 'rxjs/add/operator/shareReplay';
-import 'rxjs/add/observable/empty';
-import 'rxjs/add/observable/from';
-import 'rxjs/add/observable/timer';
-import 'rxjs/add/observable/of';
-import 'rxjs/add/observable/merge';
-import 'rxjs/add/observable/throw';
+import {
+    catchError,
+    delay,
+    exhaust,
+    filter,
+    map,
+    mapTo,
+    merge,
+    mergeMap,
+    multicast,
+    refCount,
+    retryWhen,
+    share,
+    switchMap,
+    switchMapTo,
+    take,
+    takeUntil,
+    tap
+} from "rxjs/operators";
 
 export class Client {
     private messages: Observable<IMessage>;
     private subscription: Subscription;
     private _session: Observable<WelcomeMessage>;
-    private _onClose: Observable<IMessage>;
     private challengeCallback: (challenge: Observable<any>) => Observable<string>;
     private currentRetryCount = 0;
+
+    constructor(private urlOrTransport: string | TransportInterface,
+                private realm: string,
+                private options: WampOptions = {},
+                private transport?: TransportInterface) {
+
+        this.transport = typeof urlOrTransport === 'string'
+            ? new WebSocketTransport(urlOrTransport)
+            : <TransportInterface>this.urlOrTransport;
+
+        this.subscription = new Subscription();
+
+        const open = this.transport.onOpen;
+
+        this.messages = this.transport.pipe(retryWhen((attempts: Observable<Error>) => {
+                const maxRetryDelay = 300000;
+                const initialRetryDelay = 1500;
+                const retryDelayGrowth = 1.5;
+                const maxRetries = 550;
+
+                return attempts
+                    .pipe(mergeMap((ex: any) => {
+                            console.error(ex.message);
+                            console.log('Reconnecting');
+                            const delay = Math.min(maxRetryDelay, Math.pow(retryDelayGrowth, ++this.currentRetryCount) + initialRetryDelay);
+                            return timer(Math.floor(delay));
+                        })
+                        , take(maxRetries));
+            })
+            , map((msg: IMessage) => {
+                if (msg instanceof AbortMessage) {
+                    // @todo create an exception for this
+                    asyncScheduler.schedule(() => {
+                        throw new Error('Connection ended because ' + msg.details);
+                    }, 0);
+                }
+                return msg;
+            })
+            , share());
+
+        this._onClose = this.messages.pipe(
+            filter(msg => msg instanceof AbortMessage || msg instanceof GoodbyeMessage)
+            , share());
+
+        open.pipe(
+            tap(() => {
+                this.currentRetryCount = 0;
+            })
+            , map(_ => {
+                this.options.roles = Client.roles();
+                return new HelloMessage(this.realm, this.options);
+            }))
+            .subscribe(m => this.transport.next(m));
+
+        const challengeMsg = this.messages.pipe(
+            filter((msg: IMessage) => msg instanceof ChallengeMessage)
+            , switchMap((msg: ChallengeMessage) => {
+                try {
+                    return this.challengeCallback(of(msg)).pipe(take(1));
+                } catch (e) {
+                    throw new WampChallengeException(msg);
+                }
+            })
+            , map((signature: string) => new AuthenticateMessage(signature))
+            , catchError((error: Error) => {
+                if (error instanceof WampChallengeException) {
+                    return of(error.abortMessage());
+                }
+                return throwError(error);
+            })
+            , tap(m => this.transport.next(m)));
+
+        this._session = this.messages
+            .pipe(merge(challengeMsg)
+                , filter((msg: IMessage) => msg instanceof WelcomeMessage)
+                , multicast(() => new ReplaySubject(1))
+                , refCount());
+
+        this.subscription.add(this.transport);
+    }
+
+    private _onClose: Observable<IMessage>;
+
+    get onClose(): Observable<IMessage> {
+        return this._onClose;
+    }
+
+    get onOpen(): Observable<IMessage> {
+        return this._session;
+    }
 
     private static roles() {
         return {
@@ -98,118 +173,38 @@ export class Client {
         };
     }
 
-    constructor(private urlOrTransport: string | TransportInterface,
-                private realm: string,
-                private options: WampOptions = {},
-                private transport?: TransportInterface) {
-
-        this.transport = typeof urlOrTransport === 'string'
-            ? new WebSocketTransport(urlOrTransport)
-            : <TransportInterface>this.urlOrTransport;
-
-        this.subscription = new Subscription();
-
-        const open = this.transport.onOpen;
-
-        this.messages = this.transport
-            .retryWhen((attempts: Observable<Error>) => {
-                const maxRetryDelay = 300000;
-                const initialRetryDelay = 1500;
-                const retryDelayGrowth = 1.5;
-                const maxRetries = 550;
-
-                return attempts
-                    .flatMap((ex) => {
-                        console.error(ex.message);
-                        console.log('Reconnecting');
-                        const delay = Math.min(maxRetryDelay, Math.pow(retryDelayGrowth, ++this.currentRetryCount) + initialRetryDelay);
-                        return Observable.timer(Math.floor(delay));
-                    })
-                    .take(maxRetries);
-            })
-            .map((msg: IMessage) => {
-                if (msg instanceof AbortMessage) {
-                    // @todo create an exception for this
-                    Scheduler.async.schedule(() => {
-                        throw new Error('Connection ended because ' + msg.details);
-                    }, 0);
-                }
-                return msg;
-            })
-            .share();
-
-        this._onClose = this.messages
-            .filter(msg => msg instanceof AbortMessage || msg instanceof GoodbyeMessage)
-            .share();
-
-        open
-            .do(() => {
-                this.currentRetryCount = 0;
-            })
-            .map(_ => {
-                this.options.roles = Client.roles();
-                return new HelloMessage(this.realm, this.options);
-            })
-            .subscribe(m => this.transport.next(m));
-
-        const challengeMsg = this.messages
-            .filter((msg: IMessage) => msg instanceof ChallengeMessage)
-            .switchMap((msg: ChallengeMessage) => {
-                try {
-                    return this.challengeCallback(Observable.of(msg)).take(1);
-                } catch (e) {
-                    throw new WampChallengeException(msg);
-                }
-            })
-            .map((signature: string) => new AuthenticateMessage(signature))
-            .catch((error: Error) => {
-                if (error instanceof WampChallengeException) {
-                    return Observable.of(error.abortMessage());
-                }
-                return Observable.throw(error);
-            })
-            .do(m => this.transport.next(m));
-
-        this._session = this.messages
-            .merge(challengeMsg)
-            .filter((msg: IMessage) => msg instanceof WelcomeMessage)
-            .multicast(() => new ReplaySubject(1)).refCount();
-
-        this.subscription.add(this.transport);
-    }
-
     public topic(uri: string, options?: TopicOptions): Observable<EventMessage> {
-        return this._session
-            .takeUntil(this.onClose)
-            .switchMapTo(new TopicObservable(uri, options, this.messages, this.transport));
+        return this._session.pipe(
+            takeUntil(this.onClose)
+            , switchMapTo(new TopicObservable(uri, options, this.messages, this.transport)));
     }
 
     public publish<T>(uri: string, value: Observable<T> | any, options?: PublishOptions): Subscription {
-        const obs = typeof value.subscribe === 'function' ? value as Observable<T> : Observable.of(value);
+        const obs = typeof value.subscribe === 'function' ? value as Observable<T> : of(value);
         const completed = new Subject();
 
         return this._session
-            .takeUntil(completed)
-            .takeUntil(this.onClose)
-            .mapTo(obs.do(null, null, () => {
-                completed.next(0);
-            }))
-            .exhaust()
-            .map(v => new PublishMessage(Utils.uniqueId(), options, uri, [v]))
+            .pipe(takeUntil(completed)
+                , takeUntil(this.onClose)
+                , mapTo(obs.pipe(tap(null, null, () => {
+                    completed.next(0);
+                })))
+                , exhaust()
+                , map(v => new PublishMessage(Utils.uniqueId(), options, uri, [v])))
             .subscribe(this.transport);
     }
 
     public call(uri: string, args?: Array<any>, argskw?: Object, options?: CallOptions): Observable<ResultMessage> {
-        return this._session
-            .merge(this.onClose.mapTo(Observable.throw(new Error('Connection Closed'))))
-            .take(1)
-            .switchMapTo(new CallObservable(uri, this.messages, this.transport, args, argskw, options));
+        return this._session.pipe(
+            merge(this.onClose.pipe(mapTo(throwError(new Error('Connection Closed')))))
+            , take(1)
+            , switchMapTo(new CallObservable(uri, this.messages, this.transport, args, argskw, options)));
     }
 
     public register(uri: string, callback: Function, options?: RegisterOptions): Observable<RegisteredMessage | UnregisteredMessage> {
-        return this._session
-            .merge(this.onClose.mapTo(Observable.throw(new Error('Connection Closed'))))
-            .switchMapTo(new RegisterObservable(uri, callback, this.messages, this.transport, options));
+        return this._session.pipe(
+            merge(this.onClose.pipe(mapTo(throwError(new Error('Connection Closed')))))
+            , switchMapTo(new RegisterObservable(uri, callback, this.messages, this.transport, options)));
     }
 
     public progressiveCall(uri: string, args?: Array<any>, argskw?: Object, options: CallOptions = {}): Observable<ResultMessage> {
@@ -218,28 +213,28 @@ export class Client {
         const completed = new Subject();
         const callObs = new CallObservable(uri, this.messages, this.transport, args, argskw, options);
         let retry = false;
-        return this._session
-            .merge(this.onClose.mapTo(Observable.throw(new Error('Connection Closed'))))
-            .takeUntil(completed)
-            .switchMapTo(callObs.do(null, null, () => {
+        return this._session.pipe(
+            merge(this.onClose.pipe(mapTo(throwError(new Error('Connection Closed')))))
+            , takeUntil(completed)
+            , switchMapTo(callObs.pipe(tap(null, null, () => {
                 completed.next(0);
-            }))
-            .do(() => {
+            })))
+            , tap(() => {
                 retry = false
             })
-            .retryWhen((errors: Observable<any>) => {
-                return errors
-                    .flatMap((e: WampErrorException) => {
+            , retryWhen((errors: Observable<any>) => {
+                return errors.pipe(
+                    mergeMap((e: WampErrorException) => {
                         // start retrying when we get a canceled error and continue retrying until we get a value
                         if (e.errorUri === 'wamp.error.canceled' || retry) {
                             retry = true;
-                            return Observable.of(e);
+                            return of(e);
                         }
 
-                        return Observable.empty();
+                        return empty();
                     })
-                    .delay(5000);
-            });
+                    , delay(5000));
+            }));
     }
 
     public progressiveRegister(uri: string, callback: Function, options: RegisterOptions = {}): Observable<any> {
@@ -255,14 +250,6 @@ export class Client {
 
     public close() {
         this.subscription.unsubscribe();
-    }
-
-    get onOpen(): Observable<IMessage> {
-        return this._session;
-    }
-
-    get onClose(): Observable<IMessage> {
-        return this._onClose;
     }
 }
 
