@@ -1,6 +1,4 @@
-import {Observable} from 'rxjs/Observable';
-import {Subscription} from 'rxjs/Subscription';
-import {Subscriber} from 'rxjs/Subscriber';
+import {Scheduler,Subject,Subscriber,Subscription,Observable, from, of, throwError,  } from "rxjs";
 import {WampErrorException} from '../Common/WampErrorException';
 import {ResultMessage} from '../Messages/ResultMessage';
 import {CancelMessage} from '../Messages/CancelMessage';
@@ -8,8 +6,20 @@ import {ErrorMessage} from '../Messages/ErrorMessage';
 import {CallMessage} from '../Messages/CallMessage';
 import {IMessage} from '../Messages/Message';
 import {Utils} from '../Common/Utils';
-import {Subject} from 'rxjs/Subject';
-import {Scheduler} from 'rxjs/Scheduler';
+import {
+    tap,
+    share,
+    filter,
+    mergeMap,
+    publish,
+    takeWhile,
+    finalize,
+    takeUntil,
+    take,
+    merge,
+    map,
+    refCount
+} from 'rxjs/operators';
 
 export interface CallOptions {
     receive_progress?: boolean;
@@ -32,18 +42,18 @@ export class CallObservable<ResultMsg> extends Observable<any> {
                 private options: CallOptions = {},
                 private scheduler?: Scheduler) {
         super();
-        this.messages = messages.share();
+        this.messages = messages.pipe(share());
     }
 
     public _subscribe(subscriber: Subscriber<any>): Subscription | Function | void {
         const requestId = Utils.uniqueId();
         const callMsg = new CallMessage(requestId, this.options, this.uri, this.args, this.argskw);
 
-        const msg = this.messages
-            .do(null, () => this.completed = true)
-            .filter((m: IMessage) => m instanceof ResultMessage && m.requestId === requestId)
-            .filter((m: ResultMessage) => (!!m.args || !!m.argskw) || !m.details.progress)
-            .flatMap((m: ResultMessage, index: number) => {
+        const msg = this.messages.pipe(
+            tap(null, () => this.completed = true),
+            filter((m: IMessage) => m instanceof ResultMessage && m.requestId === requestId),
+            filter((m: ResultMessage) => (!!m.args || !!m.argskw) || !m.details.progress),
+            mergeMap((m: ResultMessage, index: number) => {
                 // If there is no progress and it's the first message or there are no args, add a fake progress to
                 // the end so that the observable completes
                 if (!!m.details.progress === false && (index === 0 || (m.args || m.argskw))) {
@@ -51,28 +61,29 @@ export class CallObservable<ResultMsg> extends Observable<any> {
 
                     details.progress = true;
 
-                    return Observable.from([
+                    return from([
                         new ResultMessage(m.requestId, details, m.args, m.argskw),
                         new ResultMessage(m.requestId, {progress: false})
                     ], this.scheduler)
                 }
 
-                return Observable.of(m);
-            })
-            .publish().refCount();
+                return of(m);
+            }),
+            publish(),
+            refCount());
 
         // take until we get a message with progress: false
-        const resultMsg = msg
-            .takeWhile(m => m.details.progress || false)
-            .finally(() => this.completed = true)
-            .share();
+        const resultMsg = msg.pipe(
+            takeWhile((m:any) => m.details.progress || false),
+            finalize(() => this.completed = true),
+            share());
 
-        const error = this.messages
-            .filter((m: IMessage) => m instanceof ErrorMessage && m.errorRequestId === requestId)
-            .do(() => this.completed = true)
-            .takeUntil(msg.filter(m => !m.details.progress))
-            .flatMap((m: ErrorMessage) => Observable.throw(new WampErrorException(m.errorURI, m.args), this.scheduler))
-            .take(1);
+        const error = this.messages.pipe(
+            filter((m: IMessage) => m instanceof ErrorMessage && m.errorRequestId === requestId),
+            tap(() => this.completed = true),
+            takeUntil(msg.pipe(filter((m:any) => !m.details.progress))),
+            mergeMap((m: ErrorMessage) => throwError(new WampErrorException(m.errorURI, m.args), this.scheduler)),
+            take(1));
 
         try {
             this.webSocket.next(callMsg);
@@ -81,13 +92,11 @@ export class CallObservable<ResultMsg> extends Observable<any> {
             return;
         }
 
-        const result: Observable<ResultMessage> = error
-            .merge(resultMsg)
-            .map((m: ResultMessage) => {
+        const result: Observable<ResultMessage> = error.pipe(merge(resultMsg), map((m: ResultMessage) => {
                 const details = m.details;
                 delete details.progress;
                 return new ResultMessage(m.requestId, details, m.args, m.argskw);
-            });
+            }));
 
         const disposable = new Subscription(() => {
             if (!this.completed) {
