@@ -1,6 +1,5 @@
-import {Subject} from 'rxjs/Subject';
-import {Observable} from 'rxjs/Observable';
 import {IMessage} from '../Messages/Message';
+import {Subject,Observable} from "rxjs";
 
 export interface TransportInterface extends Subject<IMessage> {
     onOpen: Observable<Event>
